Fixed warning due to implicit conversion from Long to Double
[warn] /Users/ben/src/kojo/git/kojo/src/main/scala/net/kogics/kojo/animation/package.scala:157:35: Widening conversion from Long to Double is deprecated because it loses precision. Write `.toDouble` instead.
[warn]             val ns = nextState(s, elapsedTimeMillis)
[warn]                                   ^
[warn] one warning found

+++ src/main/scala/net/kogics/kojo/animation/package.scala
@@ -154,7 +154,7 @@ package object animation {
           }
           else {
             val elapsedTimeMillis = System.currentTimeMillis - startMillis
-            val ns = nextState(s, elapsedTimeMillis)
+            val ns = nextState(s, elapsedTimeMillis.toDouble)
             if (ns == finalState && elapsedTimeMillis > durationMillis) {
               (pic2, ns, true)
             }
